+++ ddl/db.lua
@@ -107,6 +107,10 @@ local function datetime_allowed()
     return tarantool_version_at_least(2, 10)
 end
 
+local function interval_allowed()
+    return tarantool_version_at_least(2, 10)
+end
+
 -- https://github.com/tarantool/tarantool/issues/4083
 local function transactional_ddl_allowed()
     return tarantool_version_at_least(2, 2)
@@ -155,6 +159,7 @@ return {
     multikey_path_allowed = multikey_path_allowed,
     transactional_ddl_allowed = transactional_ddl_allowed,
     datetime_allowed = datetime_allowed,
+    interval_allowed = interval_allowed,
     exclude_null_allowed = exclude_null_allowed,
 
     call_atomic = call_atomic,