+++ test/check_schema_test.lua
@@ -199,6 +199,12 @@ function g.test_datetime_index_part_type()
     end
 end
 
+function g.test_interval_index_part_type()
+    local ok, err = ddl_check.check_index_part_type('interval', 'TREE')
+    t.assert_not(ok)
+    t.assert_equals(err, "interval field type is unsupported in indexes")
+end
+
 function g.test_index_part_path()
     local index_info = {type = 'HASH'}
 
@@ -1152,6 +1158,21 @@ function g.test_field()
         t.assert_not(ok)
         t.assert_equals(err, string.format(
             [[spaces["space"].format["x"].type: datetime type ]] ..
+            [[isn't allowed in your Tarantool version (%s)]],
+            _TARANTOOL
+        ))
+    end
+
+    local ok, err = ddl_check.check_field(
+        1, {name = 'x', type = 'interval', is_nullable = false}, space_info
+    )
+    if db.v(2, 10) then
+        t.assert(ok)
+        t.assert_not(err)
+    else
+        t.assert_not(ok)
+        t.assert_equals(err, string.format(
+            [[spaces["space"].format["x"].type: interval type ]] ..
             [[isn't allowed in your Tarantool version (%s)]],
             _TARANTOOL
         ))